+++ test/unit/test_translate.py
@@ -42,25 +42,30 @@ def mock_open(*args, **kargs):
 
 @unittest.mock.patch("builtins.open", new_callable=mock_open, read_data=TEST_DATA)
 def test_translate_by_file(mock_file, mock_translator, mock_output_handler):
+    mock_translator.translate.return_value = ['', '']
+    mock_translator.batch_size = 1
     mock_file.return_value = TEST_DATA.splitlines()
     sockeye.translate.read_and_translate(translator=mock_translator, output_handler=mock_output_handler,
-                                         source='/dev/null')
+                                         chunk_size=2, source='/dev/null')
 
     # Ensure that our translator has the correct input passed to it.
     mock_translator.make_input.assert_any_call(1, "Test file line 1")
     mock_translator.make_input.assert_any_call(2, "Test file line 2")
 
-    # Ensure translate gets called twice.  Input here will be a dummy mocked result, so we'll ignore it.
-    assert mock_translator.translate.call_count == 2
+    # Ensure translate gets called once.  Input here will be a dummy mocked result, so we'll ignore it.
+    assert mock_translator.translate.call_count == 1
 
 
 @unittest.mock.patch("sys.stdin", io.StringIO(TEST_DATA))
-def test_translate_by_stdin(mock_translator, mock_output_handler):
-    sockeye.translate.read_and_translate(translator=mock_translator, output_handler=mock_output_handler)
+def test_translate_by_stdin_chunk2(mock_translator, mock_output_handler):
+    mock_translator.translate.return_value = ['', '']
+    mock_translator.batch_size = 1
+    sockeye.translate.read_and_translate(translator=mock_translator, output_handler=mock_output_handler,
+                                         chunk_size=2)
 
     # Ensure that our translator has the correct input passed to it.
     mock_translator.make_input.assert_any_call(1, "Test file line 1\n")
     mock_translator.make_input.assert_any_call(2, "Test file line 2\n")
 
-    # Ensure translate gets called twice.  Input here will be a dummy mocked result, so we'll ignore it.
-    assert mock_translator.translate.call_count == 2
+    # Ensure translate gets called once.  Input here will be a dummy mocked result, so we'll ignore it.
+    assert mock_translator.translate.call_count == 1